tidy fix

+++ src/ocl/tensorocl.cpp
@@ -140,7 +140,8 @@ void OpTensor3d(Handle& handle,
 
             network_config += std::to_string(clens[2]) + std::to_string(clens[1]) +
                               std::to_string(float_equal(miopen_beta, 0.0)) +
-                              std::to_string(blens[1] == 1) + std::to_string(max_num_wg);
+                              std::to_string(static_cast<int>(blens[1] == 1)) +
+                              std::to_string(max_num_wg);
 
             auto&& kernels = handle.GetKernels("Op2dTensorLite", network_config);
